Graded HW3
Nicely done! Loved the conciseness of your build_random_functions.
In future, comment code as you go with # instead of putting everything in the docstring if applicable. Don't print statements in long looping functions, as this greatly increases the runtime of the function when I run function as a cursory test.

Close this pull request for +1 points! **4/5**