Graded HW4
I graded this file because it was what was described in the writeup. Please clarify if necessary.

Since this was all one long script, there wasn't a lot of documentation in terms of docstrings. Some variables are unclear, in line comments are nice and clarify when present.
**Documentation: +0.75**

Again, a lot of code is repeated because of the lack of functions. There are still unnecessary code fragments floating around.
 **Style: +0.75**

A majority of the code is commented out, the final graph is generated with hardcoded variables instead of information procured from above functions.
 **Functionality: +2**

General comments:
This project turned out to be fairly difficult due to the multiple languages approach, nice job still working through the difficulties though! Ask for more help next time in figuring out how functions work and how to fix errors in your code. Make sure the final version you commit to github is how you'd want your code to run on a test computer without any modifications. I suspect everything works fine, but you forgot to uncomment certain things. You did a great job manipulating the various pattern libraries.

**Grade: 3.5/5**